requirements.txt
# Supply Chain Sentinel - Showcase Version
# SHOWCASE NOTE: This requirements file is shown for reference.
# Some dependencies may not be fully utilized in the showcase version.

crewai>=0.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
pydantic>=2.5.0
loguru>=0.7.2
feedparser>=6.0.10
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.28.0

# SHOWCASE: All dependencies are listed, but implementations are stubbed.
# The full version uses these libraries for actual execution.
